Make Everything Double

+++ BenchmarkCmp/Models/Internal/Benchmark.cs
@@ -6,7 +6,7 @@
         public double Mean { get; init; }
 
         #region Memory Diagnoser
-        public long? AllocatedBytes { get; init; }
+        public double? AllocatedBytes { get; init; }
         public double? Gen0Collects { get; init; }
         public double? Gen1Collects { get; init; }
         public double? Gen2Collects { get; init; }
@@ -24,8 +24,8 @@
         #endregion
 
         #region Native Memory Diagnoser
-        public long? NativeAllocatedBytes { get; init; }
-        public long? NativeMemoryLeak { get; init; }
+        public double? NativeAllocatedBytes { get; init; }
+        public double? NativeMemoryLeak { get; init; }
         #endregion
     }
 }
